Simulated Annealing implemented

+++ src/main/java/org/openjfx/MetaHeuristicAlgoVisualizer/FactorialArray.java
@@ -6,11 +6,11 @@ import java.util.Random;
 
 public class FactorialArray {
 	private int[] fact;
-	private int nFactorial;
+	private long nFactorial;
 	private int n;
 	Random generator = new Random();
 	
-	public int getNFactorial() {
+	public long getNFactorial() {
 		return nFactorial;
 	}
 	
@@ -25,14 +25,14 @@ public class FactorialArray {
 		nFactorial = fact[k-1];
 	}
 
- 	public ArrayList<Integer> generateIthPermutaion(int i) {
+ 	public ArrayList<Integer> generateIthPermutaion(long i) {
 		int j, k = 0;
 		Integer perm[] = new Integer[n];
 		
 		// compute factorial code
 		for (k = 0; k < n; ++k)
 		{
-			perm[k] = i / fact[n - 1 - k];
+			perm[k] = (int) (i / fact[n - 1 - k]);
 			i = i % fact[n - 1 - k];
 		}
 		
@@ -48,9 +48,9 @@ public class FactorialArray {
 	   return permList;
 	}
  	
-	public int generateNewPerm(int p1, int p2) {
+	public long generateNewPerm(long p1, long p2) {
 		double phi = generator.nextDouble()*2 - 1;
-		int newPerm = (int) (p1 + phi * (p1 - p2));
+		long newPerm = (int) (p1 + phi * (p1 - p2));
 		newPerm = Math.max(0, Math.min(nFactorial, newPerm));
 		return newPerm;
 	}
